feat: add validation functions on entities and implementation of them on Application

+++ SeniorGest/geriatricare/src/main/java/com/geriaTeam/geriatricare/applications/FuncaoApplication.java
@@ -19,15 +19,22 @@ public class FuncaoApplication {
     }
 
     public void adicionarFuncao(String nome) {
+        if (!validarNome(nome)) {
+            throw new IllegalArgumentException("Nome inválido: " + nome);
+        }
+
         FuncaoModels funcao = new FuncaoModels();
         funcao.setNome(nome);
 
-        // Adicionando a função no banco de dados
         funcaoRepository.adicionar(funcao);
     }
 
     public void removerFuncao(int id) {
-        FuncaoModels funcao = funcaoRepository.buscarPorCodigo(id);  // Busca a função pelo id
+        if (!validarId(id)) {
+            throw new IllegalArgumentException("ID inválido: " + id);
+        }
+
+        FuncaoModels funcao = funcaoRepository.buscarPorCodigo(id);
         if (funcao != null) {
             funcaoRepository.remover(id);
         } else {
@@ -36,6 +43,10 @@ public class FuncaoApplication {
     }
 
     public FuncaoModels buscarFuncao(int id) {
+        if (!validarId(id)) {
+            throw new IllegalArgumentException("ID inválido: " + id);
+        }
+
         FuncaoModels funcao = funcaoRepository.buscarPorCodigo(id);
         if (funcao == null) {
             throw new EntityNotFoundException("Função não encontrada.");
@@ -44,6 +55,13 @@ public class FuncaoApplication {
     }
 
     public void atualizarFuncao(int id, String novoNome) {
+        if (!validarId(id)) {
+            throw new IllegalArgumentException("ID inválido: " + id);
+        }
+        if (!validarNome(novoNome)) {
+            throw new IllegalArgumentException("Nome inválido: " + novoNome);
+        }
+
         FuncaoModels funcao = funcaoRepository.buscarPorCodigo(id);
         if (funcao != null) {
             funcao.setNome(novoNome);
@@ -55,5 +73,13 @@ public class FuncaoApplication {
 
     public List<FuncaoModels> buscarTodasFuncoes() {
         return funcaoRepository.buscar();
+    }
+
+    private boolean validarId(int id) {
+        return Funcao.validarId(id);
+    }
+
+    private boolean validarNome(String nome) {
+        return Funcao.validarNome(nome);
     }
 }
